Module yew_router::agent

Routing agent.

It wraps a route service and allows calls to be sent to it to update every subscriber, or just the element that made the request.

Structs

RouteAgent
The RouteAgent holds on to the RouteService singleton and mediates access to it.

RouteAgentBridge
A wrapped bridge to the route agent.

RouteAgentDispatcher
A wrapped dispatcher to the route agent.

Enums

Msg
Internal Message used for the RouteAgent.

RouteRequest
Input message type for interacting with the `RouteAgent'.