Struct yew_router::agent::RouteAgent

pub struct RouteAgent<STATE = ()> where
    STATE: RouteState,  { /* fields omitted */ }

The RouteAgent holds on to the RouteService singleton and mediates access to it.

It serves as a means to propagate messages to components interested in the state of the current route.

Warning

All routing-related components/agents/services should use the same type parameter across your application.

If you use multiple agents with different types, then the Agents won’t be able to communicate to each other and associated components may not work as intended.