Enum yew_router::matcher::CaptureVariant

pub enum CaptureVariant {
    Unnamed,
    ManyUnnamed,
    NumberedUnnamed {
        sections: usize,
    },
    Named(String),
    ManyNamed(String),
    NumberedNamed {
        sections: usize,
        name: String,
    },
}

Variants that indicate how part of a string should be captured.

Variants

Unnamed

{}

ManyUnnamed

{*}

NumberedUnnamed

{5}

Fields of NumberedUnnamed

sections: usize

Number of sections to match.

Named(String)

{name} - captures a section and adds it to the map with a given name.

ManyNamed(String)

{*:name} - captures over many sections and adds it to the map with a given name.

NumberedNamed

{2:name} - captures a fixed number of sections with a given name.

Fields of NumberedNamed

sections: usize

Number of sections to match.

name: String

The key to be entered in the Matches map.