Struct yew_router::route::Route

pub struct Route<STATE = ()> {
    pub route: String,
    pub state: STATE,
}

The representation of a route, segmented into different sections for easy access.

Fields
route: String

The route string

state: STATE

The state stored in the history api

Implementations

impl Route<()>

pub fn new_no_state<T: AsRef<str>>(route: T) -> Self

Creates a new route with no state out of a string.

This Route will have () for its state.

impl<STATE: Default> Route<STATE>

pub fn new_default_state<T: AsRef<str>>(route: T) -> Self

Creates a new route out of a string, setting the state to its default value.