Module yew_router::prelude

Prelude module that can be imported when working with the yew_router

Re-exports

pub use crate::route::RouteState;

pub use crate::service::RouteService;

pub use crate::agent::RouteAgent;

pub use crate::router::Router;

pub use crate::router::RouterState;

pub use crate::route::Route;

pub use crate::switch::Routable;

pub use crate::switch::Switch;

Structs

RouteAgentBridge

A wrapped bridge to the route agent.

RouteAgentDispatcher

A wrapped dispatcher to the route agent.

RouterAnchor

An anchor tag Component that when clicked, will navigate to the provided route.

RouterButton

Changes the route when clicked.

Type Definitions

Captures

Alias of HashMap<&'a str, String> that represent strings captured from a route.

Derive Macros

Switch

Implements the Switch trait based on attributes present on the struct or enum variants.