Module yew_router::router

Router Component.

Structs

Props

Properties for Router.

Redirect

Clonable Redirect function

Render

Owned Render function.

Router

Rendering control flow component.

Enums

Msg

Message for Router.

Traits

RedirectFn

Redirection function that takes a route that didn’t match any of the Switch variants, and converts it to a switch variant.

RenderFn

Render function that takes a switched route and converts it to HTML

RouterState

Any state that can be managed by the Router must meet the criteria of this trait.